Move 'renderedTabs' to its own component
